# api/openapi.yaml
openapi: 3.0.3
# Single source of truth for the REST contract (docs/architecture.md, section 6).
# Generated from this file:
#  - Java server interfaces + DTOs (backend, openapi-generator-maven-plugin);
#  - Swift client (apple, swift-openapi-generator) - Phase 2, client side.
#
# Scope of this version: Phase 5 (text + images, cancel/retry of a Send, renaming and deleting
# conversations, AI cost report). Draft text editing is added when a client needs it (YAGNI).
info:
 title: AI Copilot API
 version: 0.5.0
 description: Private backend for the AI Copilot iPhone and Apple Watch apps.

servers:
 - url: /

security:
 - deviceToken: []

tags:
 - name: Home
 - name: Conversations
 - name: Submissions
 - name: Drafts
 - name: Usage

paths:
 /v1/home:
  get:
   tags: [Home]
   operationId: getHome
   summary: One-call snapshot for the Watch main screen.
   description: >
    Returns the active conversation (created on first use), its current draft (created lazily),
    the latest AI request and the latest answer.
   responses:
    "200":
     description: Snapshot.
     content:
      application/json:
       schema: { $ref: "#/components/schemas/HomeSnapshot" }
    "401": { $ref: "#/components/responses/Unauthorized" }

 /v1/conversations:
  get:
   tags: [Conversations]
   operationId: listConversations
   summary: Conversations, most recently updated first.
   responses:
    "200":
     description: Conversations.
     content:
      application/json:
       schema:
        type: array
        items: { $ref: "#/components/schemas/Conversation" }
    "401": { $ref: "#/components/responses/Unauthorized" }
  post:
   tags: [Conversations]
   operationId: createConversation
   summary: Create an empty conversation and make it active on all devices.
   description: Idempotent by the client-generated id.
   requestBody:
    required: true
    content:
     application/json:
      schema: { $ref: "#/components/schemas/CreateConversationRequest" }
   responses:
    "201":
     description: Created (or already existed with this id).
     content:
      application/json:
       schema: { $ref: "#/components/schemas/Conversation" }
    "401": { $ref: "#/components/responses/Unauthorized" }

 /v1/conversations/{conversationId}:
  patch:
   tags: [Conversations]
   operationId: renameConversation
   summary: Rename a conversation. A title set by the user is never replaced by an AI-suggested one.
   parameters:
    - $ref: "#/components/parameters/ConversationId"
   requestBody:
    required: true
    content:
     application/json:
      schema: { $ref: "#/components/schemas/RenameConversationRequest" }
   responses:
    "200":
     description: Renamed.
     content:
      application/json:
       schema: { $ref: "#/components/schemas/Conversation" }
    "400":
     description: The title is empty or too long.
     content:
      application/problem+json:
       schema: { $ref: "#/components/schemas/Problem" }
    "401": { $ref: "#/components/responses/Unauthorized" }
    "404": { $ref: "#/components/responses/NotFound" }
  delete:
   tags: [Conversations]
   operationId: deleteConversation
   summary: Delete a conversation with its messages, drafts and stored images (spec 47).
   description: >
    Irreversible. Image bytes are deleted immediately. If it was the active conversation, the most recently
    updated remaining one becomes active (or a new one is created on the next home call). The AI cost report
    keeps the token counts, without any content. Idempotent: deleting a missing conversation returns 204.
   parameters:
    - $ref: "#/components/parameters/ConversationId"
   responses:
    "204": { description: Deleted (or did not exist). }
    "401": { $ref: "#/components/responses/Unauthorized" }

 /v1/active-conversation:
  put:
   tags: [Conversations]
   operationId: setActiveConversation
   summary: Make a conversation active on all devices.
   requestBody:
    required: true
    content:
     application/json:
      schema: { $ref: "#/components/schemas/SetActiveConversationRequest" }
   responses:
    "204": { description: Done. }
    "401": { $ref: "#/components/responses/Unauthorized" }
    "404": { $ref: "#/components/responses/NotFound" }

 /v1/conversations/{conversationId}/messages:
  get:
   tags: [Conversations]
   operationId: listMessages
   summary: Messages of a conversation in their stable order.
   parameters:
    - $ref: "#/components/parameters/ConversationId"
   responses:
    "200":
     description: Messages.
     content:
      application/json:
       schema:
        type: array
        items: { $ref: "#/components/schemas/Message" }
    "401": { $ref: "#/components/responses/Unauthorized" }
    "404": { $ref: "#/components/responses/NotFound" }

 /v1/conversations/{conversationId}/draft:
  get:
   tags: [Drafts]
   operationId: getCurrentDraft
   summary: The editable draft of a conversation with its attachments (created lazily).
   parameters:
    - $ref: "#/components/parameters/ConversationId"
   responses:
    "200":
     description: Current draft.
     content:
      application/json:
       schema: { $ref: "#/components/schemas/DraftDetail" }
    "401": { $ref: "#/components/responses/Unauthorized" }
    "404": { $ref: "#/components/responses/NotFound" }

 /v1/drafts/{draftId}/attachments/{attachmentId}:
  put:
   tags: [Drafts]
   operationId: registerAttachment
   summary: Register an image in a draft before its bytes are uploaded.
   description: >
    Step 1 of 2. The backend learns about the image before its bytes arrive, so a Send can wait for it
    and the Watch can show "Uploading 2/3". Idempotent by the client-generated attachment id.
    Position (order in the message) is assigned by the backend in registration order.
    Registering never starts AI inference.
   parameters:
    - $ref: "#/components/parameters/DraftId"
    - $ref: "#/components/parameters/AttachmentId"
   requestBody:
    required: true
    content:
     application/json:
      schema: { $ref: "#/components/schemas/RegisterAttachmentRequest" }
   responses:
    "200":
     description: Registered (or already registered with this id).
     content:
      application/json:
       schema: { $ref: "#/components/schemas/Attachment" }
    "401": { $ref: "#/components/responses/Unauthorized" }
    "404": { $ref: "#/components/responses/NotFound" }
    "409":
     description: The draft is no longer editable (already sent).
     content:
      application/problem+json:
       schema: { $ref: "#/components/schemas/Problem" }
  delete:
   tags: [Drafts]
   operationId: removeAttachment
   summary: Remove an image from an editable draft.
   parameters:
    - $ref: "#/components/parameters/DraftId"
    - $ref: "#/components/parameters/AttachmentId"
   responses:
    "204": { description: Removed (or did not exist). }
    "401": { $ref: "#/components/responses/Unauthorized" }
    "404": { $ref: "#/components/responses/NotFound" }
    "409":
     description: The draft is no longer editable.
     content:
      application/problem+json:
       schema: { $ref: "#/components/schemas/Problem" }

 /v1/attachments/{attachmentId}/content:
  put:
   tags: [Drafts]
   operationId: uploadAttachmentContent
   summary: Upload the bytes of a registered image (raw request body).
   description: >
    Step 2 of 2. Raw body (not multipart), so iOS background URLSession can upload it from a file.
    Idempotent: repeating the upload of the same bytes is harmless. Allowed after Send while the draft
    waits for its images. Uploading never starts AI inference by itself; it may only complete a Send
    that is already waiting for this image.
   parameters:
    - $ref: "#/components/parameters/AttachmentId"
   requestBody:
    required: true
    content:
     application/octet-stream:
      schema:
       type: string
       format: binary
   responses:
    "200":
     description: Stored and verified.
     content:
      application/json:
       schema: { $ref: "#/components/schemas/Attachment" }
    "401": { $ref: "#/components/responses/Unauthorized" }
    "404": { $ref: "#/components/responses/NotFound" }
    "409":
     description: The image can no longer be uploaded (e.g. removed or the Send was cancelled).
     content:
      application/problem+json:
       schema: { $ref: "#/components/schemas/Problem" }
    "422":
     description: Size or SHA-256 does not match the registration.
     content:
      application/problem+json:
       schema: { $ref: "#/components/schemas/Problem" }

 /v1/drafts/{draftId}/submit:
  post:
   tags: [Submissions]
   operationId: submitDraft
   summary: Explicit Send - the only operation that starts AI inference.
   description: >
    Freezes the draft and creates an AI request. Idempotent by `Idempotency-Key`: repeating the call
    with the same key returns the same request. A second submit of the same draft with a different key
    returns 409 (e.g. Watch and iPhone pressed Send at the same time).
   parameters:
    - name: draftId
      in: path
      required: true
      schema: { type: string, format: uuid }
    - name: Idempotency-Key
      in: header
      required: true
      description: Generated by the client once per user action.
      schema: { type: string, format: uuid }
   requestBody:
    required: true
    content:
     application/json:
      schema: { $ref: "#/components/schemas/SubmitDraftRequest" }
   responses:
    "202":
     description: Accepted; processing continues on the backend.
     content:
      application/json:
       schema: { $ref: "#/components/schemas/AiRequest" }
    "401": { $ref: "#/components/responses/Unauthorized" }
    "404": { $ref: "#/components/responses/NotFound" }
    "409":
     description: The draft was already submitted with another idempotency key.
     content:
      application/problem+json:
       schema: { $ref: "#/components/schemas/Problem" }
    "422":
     description: The draft has neither text nor attachments.
     content:
      application/problem+json:
       schema: { $ref: "#/components/schemas/Problem" }

 /v1/requests/{requestId}/cancel:
  post:
   tags: [Submissions]
   operationId: cancelRequest
   summary: Cancel a Send that still waits for its images; the photos and text return to the draft.
   description: >
    Allowed while the request is waitingForAttachments or blocked (no message exists yet, spec 26).
    Idempotent: cancelling a cancelled request returns it unchanged.
   parameters:
    - $ref: "#/components/parameters/RequestId"
   responses:
    "200":
     description: Cancelled.
     content:
      application/json:
       schema: { $ref: "#/components/schemas/AiRequest" }
    "401": { $ref: "#/components/responses/Unauthorized" }
    "404": { $ref: "#/components/responses/NotFound" }
    "409":
     description: Too late to cancel (the question was already sent to the AI). Code request_not_cancellable.
     content:
      application/problem+json:
       schema: { $ref: "#/components/schemas/Problem" }

 /v1/requests/{requestId}/retry:
  post:
   tags: [Submissions]
   operationId: retryRequest
   summary: Ask the AI again after a failed answer. Reuses the same question; no new message is created.
   description: >
    Allowed for a failed request whose question is still the last message of its conversation.
    Idempotent: retrying a request that is already queued, processing or completed returns it unchanged.
   parameters:
    - $ref: "#/components/parameters/RequestId"
   responses:
    "202":
     description: Queued again.
     content:
      application/json:
       schema: { $ref: "#/components/schemas/AiRequest" }
    "401": { $ref: "#/components/responses/Unauthorized" }
    "404": { $ref: "#/components/responses/NotFound" }
    "409":
     description: Cannot be retried (cancelled, still waiting for images, or newer messages exist). Code request_not_retryable.
     content:
      application/problem+json:
       schema: { $ref: "#/components/schemas/Problem" }

 /v1/requests/{requestId}:
  get:
   tags: [Submissions]
   operationId: getRequest
   summary: State of an AI request, optionally long-polling for a change.
   parameters:
    - $ref: "#/components/parameters/RequestId"
    - name: waitSeconds
      in: query
      required: false
      description: >
       0 = return immediately. Otherwise wait up to this many seconds until the state differs from
       its state at the time of the call, or the request reaches a terminal state.
      schema: { type: integer, minimum: 0, maximum: 25, default: 0 }
   responses:
    "200":
     description: Current state.
     content:
      application/json:
       schema: { $ref: "#/components/schemas/AiRequest" }
    "401": { $ref: "#/components/responses/Unauthorized" }
    "404": { $ref: "#/components/responses/NotFound" }

 /v1/usage:
  get:
   tags: [Usage]
   operationId: getUsage
   summary: AI token usage and its estimated cost for today, this month and all time.
   description: >
    Cost is estimated from token counts with the prices configured on the backend (USD per 1M tokens).
    "Today" and "this month" use the backend's reporting time zone.
   responses:
    "200":
     description: Report.
     content:
      application/json:
       schema: { $ref: "#/components/schemas/UsageReport" }
    "401": { $ref: "#/components/responses/Unauthorized" }

components:
 securitySchemes:
  deviceToken:
   type: http
   scheme: bearer
   description: Per-device token (256-bit random). Stored hashed on the server.

 parameters:
  ConversationId:
   name: conversationId
   in: path
   required: true
   schema: { type: string, format: uuid }
  DraftId:
   name: draftId
   in: path
   required: true
   schema: { type: string, format: uuid }
  RequestId:
   name: requestId
   in: path
   required: true
   schema: { type: string, format: uuid }
  AttachmentId:
   name: attachmentId
   in: path
   required: true
   description: Generated by the client; makes registration idempotent.
   schema: { type: string, format: uuid }

 responses:
  Unauthorized:
   description: Missing or invalid device token.
   content:
    application/problem+json:
     schema: { $ref: "#/components/schemas/Problem" }
  NotFound:
   description: Not found (or belongs to another user).
   content:
    application/problem+json:
     schema: { $ref: "#/components/schemas/Problem" }

 schemas:
  HomeSnapshot:
   type: object
   required: [activeConversation, draft]
   properties:
    activeConversation: { $ref: "#/components/schemas/Conversation" }
    draft: { $ref: "#/components/schemas/DraftSummary" }
    latestRequest: { $ref: "#/components/schemas/AiRequest" }
    lastAnswer: { $ref: "#/components/schemas/Message" }

  Conversation:
   type: object
   required: [id, title, mode, updatedAt]
   properties:
    id: { type: string, format: uuid }
    title: { type: string }
    mode: { $ref: "#/components/schemas/ConversationMode" }
    updatedAt: { type: string, format: date-time }

  ConversationMode:
   type: string
   description: >
    What the conversation is for, chosen at creation. danishExam adds the DU3 exam guide and the official
    study materials to the AI's instructions.
   enum: [general, danishExam]

  CreateConversationRequest:
   type: object
   required: [id]
   properties:
    id:
     type: string
     format: uuid
     description: Client-generated id; makes the call idempotent.
    mode:
     $ref: "#/components/schemas/ConversationMode"

  RenameConversationRequest:
   type: object
   required: [title]
   properties:
    title: { type: string, minLength: 1, maxLength: 80 }

  SetActiveConversationRequest:
   type: object
   required: [conversationId]
   properties:
    conversationId: { type: string, format: uuid }

  Message:
   type: object
   required: [id, conversationId, role, attachmentCount, suggestedActions, createdAt]
   properties:
    id: { type: string, format: uuid }
    conversationId: { type: string, format: uuid }
    role:
     type: string
     enum: [user, assistant]
    text:
     type: string
     nullable: true
     description: Null for image-only user messages.
    attachmentCount: { type: integer, minimum: 0 }
    suggestedActions:
     type: array
     maxItems: 3
     items: { $ref: "#/components/schemas/SuggestedAction" }
    createdAt: { type: string, format: date-time }

  SuggestedAction:
   type: object
   required: [title, prompt]
   properties:
    title: { type: string }
    prompt: { type: string }

  DraftSummary:
   type: object
   required: [id, totalAttachments, uploadedAttachments, failedAttachments]
   properties:
    id: { type: string, format: uuid }
    text: { type: string, nullable: true }
    totalAttachments: { type: integer, minimum: 0 }
    uploadedAttachments: { type: integer, minimum: 0 }
    failedAttachments: { type: integer, minimum: 0 }

  DraftDetail:
   type: object
   required: [id, conversationId, state, attachments]
   properties:
    id: { type: string, format: uuid }
    conversationId: { type: string, format: uuid }
    text: { type: string, nullable: true }
    state:
     type: string
     enum: [open, frozen, consumed]
    attachments:
     type: array
     description: In message order.
     items: { $ref: "#/components/schemas/Attachment" }

  Attachment:
   type: object
   required: [id, draftId, position, source, state, mimeType, byteSize]
   properties:
    id: { type: string, format: uuid }
    draftId: { type: string, format: uuid }
    position: { type: integer, minimum: 1 }
    source: { $ref: "#/components/schemas/AttachmentSource" }
    state:
     type: string
     enum: [pending, uploaded, failed]
    mimeType: { type: string }
    byteSize: { type: integer, format: int64 }
    failureReason: { type: string }

  AttachmentSource:
   type: string
   description: Where the image came from. Only for diagnostics/UI; processing ignores it (spec 22).
   enum: [camera, photoLibrary]

  RegisterAttachmentRequest:
   type: object
   required: [mimeType, byteSize, sha256, source]
   properties:
    mimeType:
     type: string
     enum: [image/jpeg, image/png]
    byteSize:
     type: integer
     format: int64
     minimum: 1
     maximum: 20000000
    sha256:
     type: string
     pattern: "^[0-9a-f]{64}$"
     description: Lower-case hex SHA-256 of the bytes that will be uploaded.
    source: { $ref: "#/components/schemas/AttachmentSource" }
    width: { type: integer, minimum: 1 }
    height: { type: integer, minimum: 1 }

  SubmitDraftRequest:
   type: object
   properties:
    text:
     type: string
     nullable: true
     maxLength: 4000
     description: Optional; with attachments it becomes the text of the same message.

  AiRequest:
   type: object
   required: [id, conversationId, state, createdAt]
   properties:
    id: { type: string, format: uuid }
    conversationId: { type: string, format: uuid }
    createdAt:
     type: string
     format: date-time
     description: When Send was pressed; clients show the elapsed time.
    state:
     type: string
     enum: [waitingForAttachments, blocked, queued, processing, completed, failed, cancelled]
    uploadedAttachments:
     type: integer
     description: Set in waitingForAttachments.
    totalAttachments:
     type: integer
     description: Set in waitingForAttachments.
    failedAttachments:
     type: integer
     description: Set in blocked.
    assistantMessageId:
     type: string
     format: uuid
     description: Set in completed.
    failureReason:
     type: string
     description: Set in failed; safe to show to the user.

  UsageReport:
   type: object
   required: [currency, timeZone, today, month, total]
   properties:
    currency: { type: string, description: "ISO 4217, e.g. USD." }
    timeZone: { type: string, description: "IANA zone that defines today and this month." }
    today: { $ref: "#/components/schemas/UsagePeriod" }
    month: { $ref: "#/components/schemas/UsagePeriod" }
    total: { $ref: "#/components/schemas/UsagePeriod" }

  UsagePeriod:
   type: object
   required: [answers, inputTokens, cachedInputTokens, outputTokens, cost]
   properties:
    answers: { type: integer, description: "Number of AI answers." }
    inputTokens: { type: integer, format: int64, description: "All input tokens, cached ones included." }
    cachedInputTokens: { type: integer, format: int64 }
    outputTokens: { type: integer, format: int64 }
    cost: { type: number, format: double, description: "Estimated cost in the report currency." }

  Problem:
   type: object
   description: RFC 9457 problem details.
   properties:
    type: { type: string }
    title: { type: string }
    status: { type: integer }
    detail: { type: string }
    code:
     type: string
     description: >
      Stable machine-readable code, e.g. draft_already_submitted, empty_draft, request_not_cancellable,
      request_not_retryable, invalid_title.